requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastmcp>=2.0.0
truststore>=0.8.0 ; python_version >= '3.10'
urllib3>=2.0.0 ; python_version < '3.10'
